requirements.txt
streamlit
snowflake-snowpark-python[pandas]
snowflake-connector-python
cryptography
PyJWT
requests